refactor(client-runtime): share voice input controller
