+++ packages/client-runtime/src/voice-input/controller.test.ts
@@ -3,15 +3,13 @@ import { beforeEach, describe, expect, it, vi } from "vite-plus/test";
 import {
   resetVoiceInputGlobalsForTests,
   resolveTranscriptCommit,
-  resolveVoiceComposerPresentation,
   VoiceInputController,
   VOICE_RECORDING_LIMIT_SECONDS,
   voiceInputBlocksSubmission,
-  voiceInputFreezesEditor,
   type VoiceDraftSnapshot,
   type VoiceInputControllerDependencies,
   type VoiceRecorder,
-} from "./voiceInputController";
+} from "./controller.ts";
 
 function deferred<T>() {
   let resolve!: (value: T) => void;
@@ -178,69 +176,6 @@ describe("VoiceInputController", () => {
     );
   });
 
-  it("maps voice states to stable composer actions and editor read-only state", () => {
-    expect(
-      resolveVoiceComposerPresentation({ phase: "idle", error: null, errorAction: null }, 0),
-    ).toEqual({
-      leadingAction: null,
-      trailingAction: "mic",
-      showsSend: true,
-      statusKind: null,
-      statusLabel: null,
-      confirmationEnabled: false,
-    });
-    expect(
-      resolveVoiceComposerPresentation({ phase: "preparing", error: null, errorAction: null }, 0),
-    ).toMatchObject({
-      leadingAction: "cancel",
-      trailingAction: "confirm",
-      showsSend: false,
-      statusLabel: "Preparing",
-      confirmationEnabled: false,
-    });
-    expect(
-      resolveVoiceComposerPresentation({ phase: "recording", error: null, errorAction: null }, 64),
-    ).toMatchObject({
-      leadingAction: "cancel",
-      trailingAction: "confirm",
-      showsSend: false,
-      statusLabel: "Recording 1:04",
-      confirmationEnabled: true,
-    });
-    expect(
-      resolveVoiceComposerPresentation(
-        { phase: "transcribing", error: null, errorAction: null },
-        0,
-      ),
-    ).toMatchObject({
-      statusLabel: "Transcribing",
-      confirmationEnabled: false,
-    });
-    expect(
-      resolveVoiceComposerPresentation(
-        { phase: "error", error: "Microphone unavailable", errorAction: "retry" },
-        0,
-      ),
-    ).toMatchObject({
-      leadingAction: null,
-      trailingAction: "mic",
-      showsSend: true,
-      statusKind: "error",
-      statusLabel: "Microphone unavailable",
-    });
-
-    expect(voiceInputFreezesEditor({ phase: "preparing", error: null, errorAction: null })).toBe(
-      true,
-    );
-    expect(voiceInputFreezesEditor({ phase: "recording", error: null, errorAction: null })).toBe(
-      true,
-    );
-    expect(voiceInputFreezesEditor({ phase: "transcribing", error: null, errorAction: null })).toBe(
-      true,
-    );
-    expect(voiceInputFreezesEditor({ phase: "idle", error: null, errorAction: null })).toBe(false);
-  });
-
   it("uses the native five-minute cap and commits one final transcript", async () => {
     const harness = createHarness();
     await harness.controller.start();
@@ -329,7 +264,7 @@ describe("VoiceInputController", () => {
     expect(harness.controller.currentState.error).toContain("draft changed");
   });
 
-  it("keeps the app-wide session locked until canceled native work settles", async () => {
+  it("keeps the app-wide session locked until canceled transcription work settles", async () => {
     const preparation = deferred<string>();
     const first = createHarness({ prepareTranscription: () => preparation.promise });
     const firstStart = first.controller.start();
@@ -371,7 +306,6 @@ describe("VoiceInputController", () => {
       hasError: true,
       error: "Audio route changed",
       url: "file:///voice.m4a",
-      mediaServicesDidReset: true,
     });
 
     expect(harness.commits).toEqual([]);
