+++ packages/client-runtime/src/voice-input/index.ts
@@ -1,0 +1,13 @@
+export {
+  VoiceInputController,
+  VOICE_RECORDING_LIMIT_SECONDS,
+  resolveTranscriptCommit,
+  voiceInputBlocksSubmission,
+  voiceInputFreezesEditor,
+  type VoiceDraftSnapshot,
+  type VoiceInputControllerDependencies,
+  type VoiceInputPhase,
+  type VoiceInputState,
+  type VoiceRecorder,
+  type VoiceRecorderStatus,
+} from "./controller.ts";
